% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_expl.R
\name{axis_expl}
\alias{axis_expl}
\title{Get percent of total inertia explained by principal or constrained axes}
\usage{
axis_expl(mod, axes = 1:2)
}
\arguments{
\item{mod}{cca.object}

\item{axes}{A numeric vector indicating the axes for which percent explained inertia should be returned for}
}
\value{
Returns a named vector containing explained inertia as percentages. Returns constrained inertia fo CCA and unconstrained for CA
}
\description{
Get percent of total inertia explained by principal or constrained axes
}
\author{
Mikko Vihtakari
}
