% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_plot.R
\name{axis_plot}
\alias{axis_plot}
\title{Lollipop chart of species contributions to ordination axes (ggplot2)}
\usage{
axis_plot(mod, axis = 1, sides = "both", n = 10, species = NULL,
  size_preset = "device", point_cols = c("#f8766d", "#00ba38"),
  xlab = NULL, ylab = NULL, sp_names = NULL,
  italicize_sp_names = FALSE, point_size = NULL, text_col = "white",
  map_size = FALSE, break_interval = NULL)
}
\arguments{
\item{mod}{\code{\link[vegan]{cca}} or rda object}

\item{axis}{The ordination axis for which the chart should be made for}

\item{sides}{Character vector specifying which columns should be shown. Options: \code{"both"} shows \code{n/2} from both sides of the axis ordered by species scores; \code{"contrib"} orders the species based on "contributions" to the axis (squared species scores); \code{pos} orders the species based on the species score to the axis and \code{"neg"} orders the species based on the inverted species score to the axis.}

\item{n}{Number of columns (=species) to show}

\item{species}{A character vector of species entries to be forced to the plot. The remaining of species will be taken from \code{sides} and \code{n} arguments.}

\item{size_preset}{Character. Define the size preset of the plot. Options: \code{"device"}, which is optimized for the R graph window, and \code{"pdf"}, which is optimized as subplots to go together with \code{\link{ggord_plot}}.}

\item{point_cols}{a character vector of length 2 specifying the colors for negative and positive species scores respectively.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label. If \code{NULL}, the label is generated automatically from the \code{\link[vegan]{cca}} or rda object.}

\item{sp_names}{Optional data frame specifying the species names to be used in the plot. The first column has to specify the species names used in the \code{\link[vegan]{cca}} or rda object and the second column the species names that should be used to replace the original names.}

\item{italicize_sp_names}{Logical. Should species names be italicized?}

\item{point_size}{Single number specifying the size for "lollipop heads".}

\item{text_col}{Character. Color of text for the points. The text spcifies species contributions to the axis in percentages.}

\item{map_size}{Logical. Should point size be mapped to species contributions to the axis?}

\item{break_interval}{Single number specifying the interval to be used for x-axis.}
}
\description{
Creates a lollipop chart of species contributions to ordination axes with scalable point size
}
\examples{
library(vegan)
data(dune)
axis_plot(rda(dune), break_interval = 0.5)
}
\author{
Mikko Vihtakari
}
