% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfit_table.R
\name{envfit_table}
\alias{envfit_table}
\title{Produce a data frame from an envfit object}
\usage{
envfit_table(ord, env, round_digits = 2, clean_var_names = FALSE, ...)
}
\arguments{
\item{ord}{An ordination object or other structure from which the ordination scores can be extracted (including a data frame or matrix of scores). See \code{\link[vegan]{envfit}}.}

\item{env}{Data frame, matrix or vector of environmental variables. The variables can be of mixed type (factors, continuous variables) in data frames. See \code{\link[vegan]{envfit}}.}

\item{round_digits}{integer giving the number of digits R2 and PC values should be rounded to.}

\item{clean_var_names}{logical indicating whether variable names should be tidied up. Useful for exporting the table to publications.}

\item{...}{other arguments to be passed to  \code{\link[vegan]{envfit}}}
}
\value{
Returns a data frame
}
\description{
Produces an organized data frame from an envfit object
}
\author{
Mikko Vihtakari
}
