% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{extract_species}
\alias{extract_species}
\title{Internal helper function to extract most contributing species from vegan's cca and rda objects}
\usage{
extract_species(mod, axis, sides, species, n)
}
\arguments{
\item{mod}{\code{\link[vegan]{cca}} or rda object}

\item{axis}{The ordination axis for which the extraction should be made for}

\item{sides}{Character vector specifying which columns should be shown. Options: \code{"both"} shows \code{n/2} from both sides of the axis ordered by species scores; \code{"contrib"} orders the species based on "contributions" to the axis (squared species scores); \code{pos} orders the species based on the species score to the axis and \code{"neg"} orders the species based on the inverted species score to the axis.}

\item{species}{A character vector of species entries to be forced to the plot. The remaining of species will be taken from \code{sides} and \code{n} arguments.}

\item{n}{Number of columns (=species) to show}
}
\description{
Internal helper function to extract most contributing species from vegan's cca and rda objects
}
\author{
Mikko Vihtakari
}
\keyword{internal}
