% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggord_plot.R
\name{ggord_plot}
\alias{ggord_plot}
\title{Alternative ordination plot for CCA and RDA objects (ggplot2)}
\usage{
ggord_plot(
  mod,
  env_data = NULL,
  axes = 1:2,
  size_preset = "device",
  arrow.mul = NULL,
  ordisurf_var = NULL,
  centroids = TRUE,
  envfits = NULL,
  sp_repel_method = "orditorp",
  sp_label_omit = 0.1,
  capitalize_cn = TRUE,
  site_col = "grey70",
  sp_col = "firebrick3",
  sp_cross_col = "red",
  sp_symbol,
  sp_symbol_col,
  cn_col = "darkolivegreen",
  ordi_col = "lightskyblue1",
  envfits_col = c("#82C893", "#D696C8", "#056A89", "#B5794C", "#FF9252"),
  base_size = NULL,
  site_size = NULL,
  sp_size = NULL,
  cn_size = NULL,
  ordi_size = NULL,
  envfits_size = NULL,
  main = NULL
)
}
\arguments{
\item{mod}{\code{\link[vegan]{cca}} or \code{\link[vegan]{rda}} object from vegan.}

\item{env_data}{optional data frame containing environmental data, which should be fitted using the \code{\link[vegan]{envfit}} and/or \code{\link[vegan]{ordisurf}} functions. The data frame must have the same row order and number than the species data frame used to construct \code{mod}.}

\item{axes}{numeric vector defining two ordination axes for the plot.}

\item{size_preset}{character specifying the size preset for the graphic device. Options: "device" or "pdf". Size of symbols in the plot will be automatically scaled to the device or 122 mm wide pdf (page width for many scientific journals), respectively. You can override the scaling using the \code{*_size} arguments.}

\item{arrow.mul}{factor to expand arrows in the graph. Analogous to the \code{\link[vegan]{plot.cca}} argument of similar name.}

\item{ordisurf_var}{character specifying the column name in \code{env_data} that should be used to plot \code{\link[vegan]{ordisurf}}. If \code{NULL}, no \code{\link[vegan]{ordisurf}} will be plotted.}

\item{centroids}{logical indicating whether CCA or RDA centroids should be plotted (\code{display = "cn"} in \code{\link[vegan]{plot.cca}}).}

\item{envfits}{character vector specifying the names of \code{\link[vegan]{envfit}} objects, which should be plotted.}

\item{sp_repel_method}{character specifying desired the method to avoid cluttering of species names. Options: "orditorp", "label_omit", "ggrepel" or \code{NULL}. "orditorp" uses the approach from \code{\link[vegan]{orditorp}} to select most contributing species labels, "label_omit" uses the \code{sp_label_omit} argument to omit species names, "ggrepel" uses the \code{\link[ggrepel]{geom_text_repel}} function together with the \code{sp_label_omit} argument and \code{NULL} does not omit any species names. All omited species names are illustrated using \code{+} symbols.}

\item{sp_label_omit}{numeric value giving the boundary for omiting species names. Any species labels less than this argument on either axis will be replaced by \code{+} symbols. Use \code{sp_repel_method} to disable the omiting.}

\item{capitalize_cn}{logical indicating whether centroid names should be capitalized.}

\item{site_col}{character specifying the color for sites (\code{display = "sites"} in \code{\link[vegan]{plot.cca}}).}

\item{sp_col}{character specifying the color for species (\code{display = "species"} in \code{\link[vegan]{plot.cca}}).}

\item{sp_cross_col}{character specifying the color for omited species label symbols (\code{+}).}

\item{sp_symbol}{Ignored.}

\item{sp_symbol_col}{Ignored.}

\item{cn_col}{character specifying the color for centroids (\code{display = "cn"} in \code{\link[vegan]{plot.cca}}).}

\item{ordi_col}{character specifying the color for \code{\link[vegan]{ordisurf}}.}

\item{envfits_col}{character vector specifying the colors for \code{\link[vegan]{envfit}} objects.}

\item{base_size}{base font size for ggplot2 (see \code{\link[ggplot2]{theme_bw}}).}

\item{site_size}{size parameter for site symbols}

\item{sp_size}{font size for species labels}

\item{cn_size}{font size for centroid labels}

\item{ordi_size}{cex parameter for \code{\link[vegan]{ordisurf}}.}

\item{envfits_size}{cex parameter for \code{\link[vegan]{envfit}} objects.}

\item{main}{character giving the title for the plot. See \code{\link[ggplot2]{ggtitle}}}
}
\description{
Alternative to plot vegan's CCA and RDA objects using ggplot2. Produces illustrative and good looking graphics with less code and need for customization. Plots \code{\link[vegan]{envfit}} and \code{\link[vegan]{ordisurf}} when needed.
}
\details{
All font sizes should be specified using actual font sizes (not ggplot font sizes).
}
\author{
Mikko Vihtakari with borrowed code from Jari Oksanen
}
