% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_taxa.R
\name{italicize_taxa}
\alias{italicize_taxa}
\title{Italicize species names}
\usage{
italicize_taxa(x, type = "markdown", clean_names = TRUE)
}
\arguments{
\item{x}{character vector of species names to be italicized}

\item{type}{character specifying the output type. Options: \code{"markdown"} for Rmarkdown tables, \code{"ggplot"} for ggplot axes labels and titles, and \code{"latex"} for LaTeX tables.}

\item{clean_names}{logical indicating whether periods and underscores in species names should be replaced by white space}
}
\description{
Italicizes species names in markdown and ggplot2 formats
}
\author{
Mikko Vihtakari
}
