% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_plot.R
\name{pattern_plot}
\alias{pattern_plot}
\title{Facet grid to examine patterns in multivariate data (ggplot2)}
\usage{
pattern_plot(mod, axis = 1, sides = "both", n = 10, species = NULL,
  expls = NULL, log_trans = FALSE, X = NULL, Y = NULL)
}
\arguments{
\item{mod}{\code{\link[vegan]{cca}} or \code{\link[vegan]{rda}} object from vegan.}

\item{axis}{ordination axis, which should be used for the plot}

\item{sides}{Character vector specifying which columns should be shown. Options: \code{"both"} shows \code{n/2} from both sides of the axis ordered by species scores; \code{"contrib"} orders the species based on "contributions" to the axis (squared species scores); \code{pos} orders the species based on the species score to the axis and \code{"neg"} orders the species based on the inverted species score to the axis.}

\item{n}{integer giving the number of species that should be plotted. Species are chosen based on contribution to \code{axis}}

\item{species}{character vector specifying the species names that should be forced to the plot. Any missing species will be taken from the \code{n_sps} argument. If \code{NULL} all species will be taken from the \code{n_sps} argument.}

\item{expls}{character vector specifying the environmental/explanatory variable names that should be used in the plot. If \code{NULL} and constrained model, variables will be taken from \code{mod}.}

\item{log_trans}{logical indicating whether species values should be logarithm + 1 transformed before plotting.}

\item{X}{optional species data frame. If \code{NULL}, the same data frame than in \code{mod} will be used.}

\item{Y}{optional constraining variable data frame. If \code{NULL}, the same data frame than in \code{mod} will be used.}
}
\description{
ggplot2 based facet_grid to examine patterns in multivariate data
}
\author{
Mikko Vihtakari
}
