% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ggordData.R
\name{print.ggordData}
\alias{print.ggordData}
\title{Print \code{ggordData} objects}
\usage{
\method{print}{ggordData}(x, ...)
}
\arguments{
\item{x}{\code{ggordData} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=ggord_plot]{ggordData}} objects
}
\author{
Mikko Vihtakari
}
