% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veganUtils.R
\docType{package}
\name{veganUtils}
\alias{veganUtils}
\alias{veganUtils-package}
\title{veganUtils}
\description{
Produce good looking and illustrative ordination graphics for \link[vegan]{vegan} with less code
}
\author{
\strong{Maintainer}: Mikko Vihtakari \email{mikko.vihtakari@gmail.com}

}
