% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_plot.R
\name{ord_plot}
\alias{ord_plot}
\title{Alternative ordination plot for CCA and RDA objects (base graphics)}
\usage{
ord_plot(
  mod,
  env_data = NULL,
  scaling = "species",
  size_preset = "device",
  ordisurf_var = NULL,
  centroids = TRUE,
  envfits = NULL,
  orditorp = TRUE,
  orditorp_air = 1,
  st_col = "grey70",
  st_symbol = "+",
  ordihull_st = FALSE,
  species = TRUE,
  sp_col = "black",
  sp_symbol = 20,
  sp_symbol_col = "black",
  sp_font = 3,
  latin_sp_names = FALSE,
  cn_col = "darkolivegreen",
  ordi_col = c("lightskyblue1", "lightsalmon1"),
  envfits_col = c("#649971", "#A47399", "#03617E", "#B5794C", "#FF9252", "#FF5F68",
    "#449BCF"),
  main_cex = NULL,
  st_cex = NULL,
  sp_cex = NULL,
  sp_symbol_cex = NULL,
  cn_cex = NULL,
  ordi_cex = NULL,
  envfits_cex = NULL,
  main = NULL,
  axis_labs = TRUE,
  legend_pos = "topleft",
  title.main = NULL,
  title.sub = NULL,
  ...
)
}
\arguments{
\item{mod}{\code{\link[vegan]{cca}} or \code{\link[vegan]{rda}} object from vegan.}

\item{env_data}{optional data frame containing environmental data, which should be fitted using the \code{\link[vegan]{envfit}} and/or \code{\link[vegan]{ordisurf}} functions. The data frame must have the same row order and number than the species data frame used to construct \code{mod}.}

\item{scaling}{Scaling for species and site scores. See the \code{scaling} argument in \code{\link[vegan]{plot.cca}}).}

\item{size_preset}{character specifying the size preset for the graphic device. Options: "device" or "pdf". Size of symbols in the plot will be automatically scaled to the device or 122 mm wide pdf (page width for many scientific journals), respectively. You can override the scaling using the \code{*_cex} arguments.}

\item{ordisurf_var}{character specifying the column name in \code{env_data} that should be used to plot \code{\link[vegan]{ordisurf}}. If \code{NULL}, no \code{\link[vegan]{ordisurf}} will be plotted.}

\item{centroids}{logical indicating whether CCA or RDA centroids should be plotted (\code{display = "cn"} in \code{\link[vegan]{plot.cca}}).}

\item{envfits}{character vector specifying column names in \code{env_data}, which should be ploted with the \code{\link[vegan]{envfit}} function.}

\item{orditorp}{logical indicating whether \code{\link[vegan]{orditorp}} should be used to avoid cluttered species names.}

\item{orditorp_air}{numeric defining the air argument for \code{\link[vegan]{orditorp}}.}

\item{st_col}{character specifying the color for sites (\code{display = "sites"} in \code{\link[vegan]{plot.cca}}). Alternatively column name in \code{env_data} and site color will be scaled to levels in the variable. Colors will be taken from the \code{envfits_col} argument.}

\item{st_symbol, sp_symbol}{\code{\link[graphics]{pch}} code for site and species symbols, respectively.}

\item{ordihull_st}{logical indicating whether \code{\link{ordihull}} should be drawn around sites if site colors are scaled to \code{env_data} (see \code{st_col}).}

\item{sp_col}{character specifying the color for species (\code{display = "species"} in \code{\link[vegan]{plot.cca}}).}

\item{sp_font}{An integer which specifies which font to use for text. If possible, device drivers arrange so that 1 corresponds to plain text (the default), 2 to bold face, 3 to italic and 4 to bold italic. Also, font 5 is expected to be the symbol font, in Adobe symbol encoding. On some devices font families can be selected by family to choose different sets of 5 fonts. From \code{\link[graphics]{par}}.}

\item{latin_sp_names}{logical indicating whether species abbreviations should be converted to latin names. This functionality is probably not working. Check.}

\item{cn_col}{character specifying the color for centroids (\code{display = "cn"} in \code{\link[vegan]{plot.cca}}).}

\item{ordi_col}{character specifying the color for \code{\link[vegan]{ordisurf}}.}

\item{envfits_col}{character vector specifying the colors for \code{\link[vegan]{envfit}} objects.}

\item{main_cex}{cex parameter for axes titles (\code{\link[graphics]{par}(cex)}).}

\item{st_cex}{cex parameter for site symbols.}

\item{sp_cex, sp_symbol_cex}{cex parameter for species names and symbols, respectively.}

\item{cn_cex}{cex parameter for centroids.}

\item{ordi_cex}{cex parameter for \code{\link[vegan]{ordisurf}}.}

\item{envfits_cex}{cex parameter for \code{\link[vegan]{envfit}} objects.}

\item{main}{character specifying title for the plot. Overlaps with title.main. Fix.}

\item{legend_pos}{character specifying the location of legend (only added if colors are mapped to variables). Set to \code{NULL} to remove the legend.}

\item{title.main}{Overlaps with main. Fix.}

\item{title.sub}{character specifying sub-title for the plot.}

\item{...}{parameters passed to \code{\link[vegan]{plot.cca}}}
}
\description{
Alternative to plot vegan's CCA and RDA objects producing illustrative and good looking graphics with less code and need for customization. Plots \code{\link[vegan]{envfit}} and \code{\link[vegan]{ordisurf}} when needed.
}
\details{
While the function produces OK looking plots at the moment (and they seem to be correct too), the function needs to be rewritten to comply with vegan's "dialect". Also some features are redundant. I will do this rewriting at some point in the future when I have time. Also add examples.
}
\author{
Mikko Vihtakari
}
